Approximate monotonic spline on data

It aproximates data with monotonic spline function by fitting GAM model.

approx_with_monotonic_spline(bb_response_data, response_var, pred_var,
  env = parent.frame(), increasing, ...)

Arguments

response_var
Name of response value from bb_response_data.

pred_var
Name of predictor value from bb_response_data.

env
Formula environment that should be used for fitting gam model.

increasing
If TRUE, spline approximation is increasing, if FALSE decreasing.

...
Other arguments passed to mgcv::s function.

bb_response_data.
Blackbox response data, for example pdp curve.

Value

Object of class "gam". See mgcv::gamObject

Examples

x <- sort(rnorm(20, 5, 5))
y <- rnorm(20, 2, 2)
env <- new.env()
approx_with_monotonic_spline(data.frame(x = x, y = y), "y", "x", env, TRUE)

#> Error in approx_with_monotonic_spline(data.frame(x = x, y = y), "y", "x",     env, TRUE): could not find function "approx_with_monotonic_spline"