Plot method for 'xspliner' model

The method provides all plotting methods offered by 'xspliner' package. See plot_variable_transition and plot_model_comparison for more details.

# S3 method for xspliner
plot(x, variable_names = NULL, model = NULL,
  plot_response = TRUE, plot_approx = TRUE, data = NULL,
  plot_data = FALSE, plot_deriv = FALSE, n_plots = 6,
  sort_by = NULL, use_coeff = TRUE, compare_with = list(),
  prediction_funs = list(function(object, newdata) predict(object,
  newdata)), ...)

Arguments

x
Object of class 'xspliner'

variable_names
Names of predictors which transitions should be plotted.

model
Base model that xspliner is based on.

plot_response
If TRUE black box model response is drawn.

plot_approx
If TRUE black box model response approximation is drawn.

data
Training data used for building x model. Required for plot_data option and model comparing.

plot_data
If TRUE raw data is drawn.

plot_deriv
If TRUE derivative of approximation is showed on plot.

n_plots
Threshold for number of plots when plotting all variables.

sort_by
When comparing models determines according to which model should observations be ordered.

use_coeff
If TRUE both PDP function and its approximation is scaled with corresponding surrogate model coefficient.

compare_with
Named list. Other models that should be compared with xspliner and model.

prediction_funs
Prediction functions that should be used in model comparison.

...
Another arguments passed into model specific method.