Builds predictive model based GLM.

See details in vignette("xspliner").

xspline(object, ...)

# S3 method for default
xspline(object, lhs = NULL, response = NULL,
  predictors = NULL, data = NULL, form = "additive", bare = NULL,
  env = parent.frame(), ...)

# S3 method for formula
xspline(object, model, data = NULL,
  consider = "specials", env = parent.frame(), ...)

# S3 method for explainer
xspline(object, env = parent.frame(), ...)

Arguments

object
Predictive model, formula or explainer (see DALEX) object.

...
Other arguments passed to xspline methods or build_xspliner.

lhs
Left-hand side of model formula. Can be transformed response.

response
Name of response variable of model.

predictors
Predictor values that should be used in final model.

data
Training data of model.

form
Can be 'additive' (default) or 'multiplicative'. Specifies formula form in final model.

bare
Variable names that mustn't be transformed in final model.

env
Environment in which optional variables passed into parameters are stored. variables transformation. See vignette("xspliner") for details.

model
When object is formula - predictive model. Basic model used for extracting predictors transformation.

consider
One of c("specials", "all"). If "specials", only components with xs or xf call are considered in transition.

Value

GLM object of class 'xspliner'.